+++ extensions/fathom-analytics-stats/CHANGELOG.md
@@ -1,3 +1,8 @@
 # Fathom Analytics Changelog
 
+## Updated icon and shortcut - {PR_MERGE_DATE}
+
+- Updated menu bar icon
+- Added shortcut to open in Fathom
+
 ## [Initial Version] - 2024-01-20